implementation

+++ src/types/PositionPlannerTypes.sol
@@ -12,14 +12,14 @@ struct Plan {
 struct PositionDefinition {
     int24 offsetLower; // tick offset from the current tick for the position's lower bound
     int24 offsetUpper; // tick offset from the current tick for the position's upper bound
-    uint24 weight; // allocation weight in mps (1e7 = 100%); all weights in a plan must sum to 1e7
+    uint24 weight; // allocation weight in mps (1e7 = 100%); explicit weights in a plan must sum to <= 1e7
 }
 
 /// @notice Generic struct representing a liquidity position
 struct Position {
-    uint128 amount0;
-    uint128 amount1;
+    uint256 amount0;
+    uint256 amount1;
     int24 tickLower;
     int24 tickUpper;
-    uint128 liquidity;
+    uint256 liquidity;
 }
